Refactor SchemaToPHP helper
